requirements.txt
paddlepaddle>=1.6.1
parl==1.3.1
rlschool==0.3.1
gym